refactor(snackbar): move useSnackbar hook to hooks directory
- Moved useSnackbar to own file useSnackbar.ts in hooks directory
