+++ src/components/SnackbarContext.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext } from 'react';
+import { createContext } from 'react';
 import { AlertColor } from '@mui/material';
 
 interface SnackbarContextType {
@@ -6,12 +6,3 @@ interface SnackbarContextType {
 }
 
 export const SnackbarContext = createContext<SnackbarContextType | null>(null);
-
-export const useSnackbar = () => {
-    const context = useContext(SnackbarContext);
-    if (!context) {
-        throw new Error('useSnackbar must be used within a SnackbarProvider');
-    }
-
-    return context;
-};
